feat: move constants into utils folder, remove unused lib
