--- libs/constants/README.md
@@ -1,7 +1,0 @@
-# constants
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test constants` to execute the unit tests via [Jest](https://jestjs.io).
