--- libs/constants/jest.config.ts
@@ -1,16 +1,0 @@
-/* eslint-disable */
-module.exports = {
-  displayName: "constants",
-  preset: "../../jest.preset.js",
-  globals: {
-    "ts-jest": {
-      tsconfig: "<rootDir>/tsconfig.spec.json",
-    },
-  },
-  testEnvironment: "node",
-  transform: {
-    "^.+\\.[t|j]sx?$": ["babel-jest", { presets: ["@nrwl/react/babel"] }],
-  },
-  moduleFileExtensions: ["ts", "tsx", "js", "jsx"],
-  coverageDirectory: "../../coverage/libs/constants",
-};
