+++ libs/utils/src/lib/stringifyDate.ts
@@ -1,5 +1,5 @@
-import { DATE_FORMAT } from "@routine-support/constants";
 import { Moment } from "moment";
+import { DATE_FORMAT } from "./constants";
 
 export const stringifyDate = (date: Moment): string => {
   return date.format(DATE_FORMAT);
