+++ libs/utils/src/lib/stringifyTime.ts
@@ -1,5 +1,5 @@
-import { TIME_FORMAT } from "@routine-support/constants";
 import { Moment } from "moment";
+import { TIME_FORMAT } from "./constants";
 
 export const stringifyTime = (time: Moment): string => {
   return time.format(TIME_FORMAT);
