+++ tsconfig.base.json
@@ -17,7 +17,6 @@
     "baseUrl": ".",
     "paths": {
       "@routine-support/api-client": ["libs/api-client/src/index.ts"],
-      "@routine-support/constants": ["libs/constants/src/index.ts"],
       "@routine-support/domains": ["libs/domains/src/index.ts"],
       "@routine-support/forms": ["libs/forms/src/index.ts"],
       "@routine-support/pictograms": ["libs/pictograms/src/index.ts"],
